+++ sql/posts_updates.sql
@@ -1,8 +1,8 @@
 CREATE TABLE posts_updates (
 	post_id VARCHAR(6),
-	score INTEGER(6),
+	score NUMERIC,
 	ratio DECIMAL(3, 2),
-	num_comments INTEGER(10),
-	timestamp_update INTEGER(11)
+	num_comments NUMERIC,
+	timestamp_update NUMERIC
 );
